Type assignement and inference.
Inference is when TypeScript automatically deducts type of a variable or constant
without code explicidly assigning one. Once it does that, said type is enforced from that
point on just as if it was assigned explicidly.

This happens when you assign variable a value on creation, for example:

let z = 5;
z = 'ok'

Will result in TypeScript error. Since z was initialized with a value of 5, type number was
automatically assigned, and therefore trying to put a string in it is rejected and throws an error.

Were z to be declared with const, rather than let, it would also enforce that it's value is 5 and anything
else would also throw an error. Than again vanilla js would do that too since const is not a TypeScript
keyword, but part of vanilla js.

If you however do this:

let z;
z = 10;
z= 'ok'

No errors will occur. Here TypeScript too infered a type on variable z when it was declared.
The type "any". Since there was no initialization at the moment of it's declaration, the value of
this variable was unknown and therefore given the most generic type of "any".

With that type in tow, both a number and string values are acceptable and no TypeScript error occures.
That being said the very namesake of the language is types and "any" is basically "no type" and
a sign of things gone wrong. You don't employ TypeScript to just plain avoid it's core purpose - that
of enforcing types and better code quality that follows.